bug fixes: use FloatTensor not LongTensor

+++ zz_scripts/dae_train/compress.py
@@ -58,12 +58,12 @@ def compress(state_dict_path, train_file_path, num_symptoms, output_dir):
     end = start + batch_size
     with torch.no_grad():
         if num_samples <= batch_size:
-            tensor = torch.LongTensor(race_symptoms.todense())
+            tensor = torch.FloatTensor(race_symptoms.todense())
             compressed = dae.encoder(to_device(tensor, device)).numpy()
         else:
             compressed = np.zeros((num_samples, target_dim), dtype=np.float32)
             while end <= num_samples:
-                tensor = torch.LongTensor(race_symptoms[start: end, :].todense())
+                tensor = torch.FloatTensor(race_symptoms[start: end, :].todense())
                 temp = dae.encoder(to_device(tensor, device)).numpy()
                 compressed[start:end, :] = temp
